UpdatedExchange

+++ Pg.71 number 7/Pg.71 number 7/Number71.cpp
@@ -7,15 +7,15 @@ using std::cin;
 
 int main()
 {
-	int HowMuchMoney;
+	double HowMuchMoney;
 
 
 
 	std::cout << "Please enter the amount of U.S Dollars you are exchanging" << std::endl;
 	std:cin >> HowMuchMoney;
-	int MexicanMoney = HowMuchMoney * 12.8863;
-	int JapanMoney = HowMuchMoney * 82.34;
-	int BritMoney = HowMuchMoney / .6318;
+	double MexicanMoney = HowMuchMoney * 12.8863;
+	double JapanMoney = HowMuchMoney * 82.34;
+	double BritMoney = HowMuchMoney / .6318;
 	std::cout << "\n" << MexicanMoney << " Pesos" << std::endl;
 	std::cout << "\n" << JapanMoney << " Yen" << std::endl;
 	std::cout << "\n" << BritMoney << " Pounds" << std::endl;
